♻️ Refactor: compress export default statments from the same source 'fixtures'
