+++ fixtures/index.js
@@ -1,0 +1,3 @@
+export { default as regions } from './regions';
+export { default as categories } from './categories';
+export { default as restaurants } from './restaurants';
